Move parser from spock to util
